Added new controller
